api-backend/pyproject.toml
[tool.poetry]
name = "api-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2 = "^2.9.3"
django-model-utils = "^4.2.0"
python-dotenv = "^0.20.0"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
natasha = "^1.4.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
